requirements.txt
numpy==1.25.0
matplotlib==3.8.0
Pillow==10.1.0
tqdm==4.66.1
torch==2.1.0
torchvision==0.16.0
nltk==3.8.1
